backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai==0.3.1
openai==1.3.5
anthropic==0.7.1
python-dotenv==1.0.0
loguru==0.7.2
SpeechRecognition==3.10.0
PyAudio==0.2.14
ffmpeg-python==0.2.0
